+++ AzzyBot-Next/Utilities/Helpers/ApiEndpoints.cs
@@ -1,8 +1,9 @@
-namespace AzzyBot.Utilities.Records.AzuraCast;
+namespace AzzyBot.Utilities.Helpers;
 
 public static class ApiEndpoints
 {
     public const string Admin = "admin";
+    public const string Files = "files";
     public const string NowPlaying = "nowplaying";
     public const string Playlist = "playlist";
     public const string Playlists = "playlists";
